Abstracted cnetstats.
